rename wrapper to indicate it is a pymc model wrapper
