+++ pyloo/wrapper/pymc_wrapper.py
